requirements.txt
asgiref==3.4.1
attrs==21.4.0
cfgv==3.3.1
distlib==0.3.4
Django==4.0.1
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==3.2.1
Faker==11.3.0
filelock==3.4.2
identify==2.4.4
iniconfig==1.1.1
nodeenv==1.6.0
packaging==21.3
platformdirs==2.4.1
pluggy==1.0.0
pre-commit==2.17.0
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
virtualenv==20.13.0
